Disable xqlint when running xqsuite tests
bugs in xqlint were leading to parsing errors, which were preventing xqsuite tests from running. this commit works around the problem by disabling the xqlint checks that were being performed when the run test command was fired. we lose 2 other checks that were being performed - 1. making sure that the open tab is a library module and 2. making sure that it contains xqsuite tests. so not a fix by any means.