Fix errors when ?info=/db/... endpoint is called
This endpoint is called in various scenarios, and they don't all involve downloading xars. When the ?info endpoint is called (e.g., deployment.xql?info=%2Fdb%2Fapps%2Fdoc), eXide 3.1.0 would raise a 400 error, because, since this request has nothing to do with serialization, indent and expand-xincludes parameters are not supplied. The values chosen for the defaults here - indent=no and expand-xincludes=no - simply match that of the pre-eXide 3.1.0 defaults.